US193682: Refactor integration tests listeners and grouping to remove dependency on cloud foundry packages
